+++ server/utils/bcrypt-helper.js
@@ -1,0 +1,7 @@
+import bcrypt from "bcrypt";
+
+export default {
+  async validatePassword (password, userHashedPassword) {
+    await bcrypt.compare(password, userHashedPassword);
+  }
+}
